Multiple fixes, documentations and logging updates

+++ bot/bdd/classes/associationsManager.js
@@ -1,9 +1,24 @@
 const { Association } = require('./association.js');
 
+/**
+ * AssociationsManager class
+ * @class
+ * @classdesc Manage the associations
+ * @alias AssociationsManager
+ * @static
+ */
 class AssociationsManager
 {
+    /**
+     * Get an association by its name
+     * @param {string} name - The association's name (will be reconstructed by removing spaces and "ISEN" from the name) 
+     * @returns {Promise<Association|null>} The association
+     * @async
+     */
     static async getAssociation(name)
     {
+        if(typeof name !== "string")
+            throw new Error("The name must be a string");
         name = name.toLowerCase();
         name = name.replace(/ISEN/g, "").replace(/ /g, "");
         const connection = global.sqlConnection;
@@ -14,8 +29,16 @@ class AssociationsManager
             return new Association(data.asso_id);
     }
 
+    /**
+     * Get an association by its ID
+     * @param {number|string} id - The association's ID
+     * @returns {Promise<Association|null>} The association
+     * @async
+     */
     static async getAssociationByID(asso_id)
     {
+        if(typeof asso_id !== "number" && typeof asso_id !== "string")
+            throw new Error("The ID must be a number or a string");
         const connection = global.sqlConnection;
         const query = "SELECT * FROM association WHERE asso_id = ?";
         const data = await connection(query, [asso_id]);
@@ -24,6 +47,11 @@ class AssociationsManager
             return new Association(asso_id);
     }
 
+    /**
+     * Get all the associations
+     * @returns {Promise<Array<Association>>} The associations
+     * @async
+     */
     static async getAssociations()
     {
         const connection = global.sqlConnection;
@@ -39,9 +67,22 @@ class AssociationsManager
         }
     }
 
+    /**
+     * add an association
+     * @param {object} association - The association's data
+     * @param {string} association.name - The association's name
+     * @param {string} association.description - The association's description
+     * @param {string} association.icon - The association's icon
+     * @returns {Promise<Association|null>} The association
+     * @async
+     */
     static async addAssociation(association)
     {
-        if (await this.getAssociation(association.name) !== null) return;
+        if (typeof association !== "object")
+            throw new Error("The association must be an object");
+        if (typeof association.name !== "string")
+            throw new Error("The association's name must be a string");
+        if (await this.getAssociation(association.name) !== null) return null;
         const connection = global.sqlConnection;
         const query = "INSERT INTO association (name, description, icon) VALUES (?, ?, ?)";
         const values = [association.name, association.description, association.icon];
@@ -49,8 +90,16 @@ class AssociationsManager
         return await this.getAssociation(association.name);
     }
 
+    /**
+     * Delete an association
+     * @param {string|number} association - The association's Id!
+     * @returns {Promise<void>}
+     * @async
+    */
     static async deleteAssociation(asso_id)
     {
+        if (typeof asso_id !== "number" && typeof asso_id !== "string")
+            throw new Error("The association's ID must be a number or a string");
         if (await this.getAssociationByID(asso_id) === null) return;
         const connection = global.sqlConnection;
         const query = "DELETE FROM association WHERE asso_id = ?";
